==> 4_cell_table_annotation.ipynb <==
### Adding stuff to the cell table for downstream visualization and such

# import needed libraries
import os
import numpy as np
import pandas as pd

os.getcwd()

os.chdir('/krummellab/data1/zachsten')
os.getcwd()

base_dir = base_dir = "personal/MIBI_data_files/zach_test_sets/"

cell_table_path = os.path.join(base_dir, 'segmentation_v1/cell_table_cell_rem/cell_table_arcsinh_transformed.csv')
cell_table = pd.read_csv(cell_table_path)
cell_table[

cell_table['fov_info'] = None
def map_fov_to_description(fov_value):
    # Find the substring after "Set"
    after_set = fov_value.split("Set")[-1]  # This splits the string on "Set" and takes the last part
    
    # Now apply your mapping logic on 'after_set' instead of the whole 'fov_value'
    if 'a' in after_set:
        return 'tumor'
    elif 'd' in after_set:
        return 'uiLN_follicle'
    elif 'e' in after_set:
        return 'uiLN_padra'
    elif 'f' in after_set:
        return 'metLN_tumor'
    elif 'g' in after_set:
        return 'metLN_follicle'
    elif 'h' in after_set:
        return 'metLN_para'
    else:
        return 'unknown'  # Handle cases where the character is not present

# Assuming 'cell_table' and its 'fov' column exist
# Apply the modified function to create the "fov_info" column
cell_table['fov_info'] = cell_table['fov'].apply(map_fov_to_description)

# Display the updated DataFrame to see the 'fov_info' column
print(cell_table[['fov', 'fov_info']])

#Fill in where 0 is going to be our control sample.
cell_table['patient_ID'] = None
def extract_patient_id(fov_value):
    # Regular expression to match the first sequence of digits (\d+) immediately after a hyphen (-)
    match = re.search(r'-(\d+)', fov_value)
    if match:
        # Extract and return the patient ID as an integer
        return int(match.group(1))
    else:
        # Return None or an appropriate placeholder if no patient ID is found
        return None
# Apply the custom function to create the "fov_info" column
cell_table['patient_ID'] = cell_table['fov'].apply(extract_patient_id)
cell_table['patient_ID'] = cell_table['patient_ID'].fillna(0).astype(int)
cell_table['patient_ID'].unique()

cell_table['fov_size'] = None
# Define a custom function to map values in "fov" to descriptions
def map_fov_to_description(fov_value):
    if '800um' in fov_value:
        return '800um'
    else:
        return '400um'  # Handle cases where neither 'e' nor 'f' is present

# Apply the custom function to create the "fov_info" column
cell_table['fov_size'] = cell_table['fov'].apply(map_fov_to_description)

# Display the updated DataFrame
print(cell_table['fov_size'].unique())

cell_table.columns

cell_table.to_csv(os.path.join(base_dir, 'segmentation_v1/cell_table_annotated/cell_table_annotated_arcsinh_transformed.csv'))

## Can add in more stuff later but for right now let's just keep it going.
